+++ src/lib/converter/context.ts
@@ -1,5 +1,5 @@
 import { ok as assert } from "assert";
-import * as ts from "typescript";
+import ts from "typescript";
 
 import {
     Reflection,