+++ src/lib/converter/factories/signature.ts
@@ -1,5 +1,5 @@
-import * as ts from "typescript";
-import * as assert from "assert";
+import ts from "typescript";
+import assert from "assert";
 import {
     ConversionFlags,
     DeclarationReflection,